[REF] Changed for Optional and modified exceptions

+++ src/main/java/com/excilys/computerdatabase/dto/CompanyDTO.java
@@ -2,23 +2,23 @@ package main.java.com.excilys.computerdatabase.dto;
 
 public class CompanyDTO {
 
-    private int id;
+    private Long id;
     private String name;
 
     public CompanyDTO() {
     }
 
-    public CompanyDTO(int id, String name) {
+    public CompanyDTO(Long id, String name) {
         super();
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
