+++ src/main/java/com/excilys/computerdatabase/mapper/ComputerMapper.java
@@ -25,14 +25,14 @@ public enum ComputerMapper {
      */
     public Computer createComputer(ResultSet rs) throws SQLException {
 
-        int id = rs.getInt("id");
+        Long id = rs.getLong("id");
         String name = rs.getString("cpuname");
         Date introducedDate = rs.getDate("introduced");
         LocalDate introduced = (introducedDate == null) ? null : introducedDate.toLocalDate();
         Date discontinuedDate = rs.getDate("discontinued");
         LocalDate discontinued = (discontinuedDate == null) ? null : discontinuedDate.toLocalDate();
         String companyName = rs.getString("companyname");
-        int companyId = rs.getInt("companyid");
+        Long companyId = rs.getLong("companyid");
         Company company = null;
         if (companyId != 0) {
             company = new Company.Builder(companyId).withName(companyName).build();
@@ -58,7 +58,7 @@ public enum ComputerMapper {
 
         LocalDate introduced = null;
         LocalDate discontinued = null;
-        int companyId;
+        Long companyId;
 
         DateTimeFormatter format = DateTimeFormatter.ofPattern("yyyy-MM-dd");
 
@@ -86,7 +86,7 @@ public enum ComputerMapper {
             throw new IllegalArgumentException();
         } else {
             try {
-                companyId = Integer.parseInt(pcompanyId);
+                companyId = Long.parseLong(pcompanyId);
             } catch (NumberFormatException e) {
                 throw e;
             }
@@ -117,8 +117,8 @@ public enum ComputerMapper {
 
         LocalDate introduced = null;
         LocalDate discontinued = null;
-        int companyId;
-        int id;
+        Long companyId;
+        Long id;
 
         DateTimeFormatter format = DateTimeFormatter.ofPattern("yyyy-MM-dd");
 
@@ -146,7 +146,7 @@ public enum ComputerMapper {
             throw new IllegalArgumentException();
         } else {
             try {
-                companyId = Integer.parseInt(pcompanyId);
+                companyId = Long.parseLong(pcompanyId);
             } catch (NumberFormatException e) {
                 throw e;
             }
@@ -156,7 +156,7 @@ public enum ComputerMapper {
             throw new IllegalArgumentException();
         } else {
             try {
-                id = Integer.parseInt(pid);
+                id = Long.parseLong(pid);
             } catch (NumberFormatException e) {
                 throw e;
             }
